Refactoring ApiTestCase
