+++ tests/SwaggerTestCaseTest.php
@@ -2,14 +2,14 @@
 
 namespace Test;
 
+use ByJG\ApiTools\Base\Schema;
+
 class SwaggerTestCaseTest extends TestingTestCase
 {
 
     public function setUp()
     {
-        $this->filePath = __DIR__ . '/rest/swagger.json';
-
-        // This is important!
-        parent::setUp();
+        $schema = Schema::getInstance(file_get_contents(__DIR__ . '/rest/swagger.json'));
+        $this->setSchema($schema);
     }
 }
